+++ wifipumpkin3/modules/__init__.py
@@ -2,6 +2,7 @@ import os
 import pkgutil
 import sys
 import importlib
+import re
 from wifipumpkin3.core.utility.printer import display_messages
 
 # https://stackoverflow.com/questions/3365740/how-to-import-all-submodules
@@ -15,9 +16,11 @@ def import_submodules(package, recursive=True):
     if isinstance(package, str):
         package = importlib.import_module(package)
     results = {}
+    parser = re.compile("(\s*)wifipumpkin3.modules.(\s*)")
     for loader, name, is_pkg in pkgutil.walk_packages(package.__path__):
         full_name = package.__name__ + "." + name
-        results[full_name] = importlib.import_module(full_name)
+        short_name = "{}.{}".format(parser.sub("", package.__name__), name)
+        results[short_name] = importlib.import_module(full_name)
         if recursive and is_pkg:
             results.update(import_submodules(full_name))
     return results